# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "教务系统",
        "description": "教务系统接口文档",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
...
        "/api/teacher/course_list": {
            "get": {
                "tags": [
                    "课程管理"
                ],
                "summary": "课程列表",
                "operationId": "4a8cbc03cdddfa644fbc8e017766bdf5",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页数"
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "每页数量"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "返回码",
                                            "type": "integer"
                                        },
                                        "message": {
                                            "description": "错误信息",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "返回数据",
                                            "properties": {
                                                "meta": {
                                                    "description": "元信息",
                                                    "properties": {
                                                        "count": {
                                                            "description": "当前页的项目数",
                                                            "type": "integer"
                                                        },
                                                        "perPage": {
                                                            "description": "每页显示的项目数",
                                                            "type": "integer"
                                                        },
                                                        "currentPage": {
                                                            "description": "当前页码",
                                                            "type": "integer"
                                                        },
                                                        "lastPage": {
                                                            "description": "最后一页的页码",
                                                            "type": "integer"
                                                        },
                                                        "total": {
                                                            "description": "总数",
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "list": {
                                                    "description": "数据列表",
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "description": "唯一标识",
                                                                "type": "integer"
                                                            },
                                                            "date": {
                                                                "description": "日期",
                                                                "type": "string"
                                                            },
                                                            "course_name": {
                                                                "description": "课程名",
                                                                "type": "string"
                                                            },
                                                            "cost": {
                                                                "description": "费用",
                                                                "type": "string"
                                                            },
                                                            "bill_id": {
                                                                "description": "账单id",
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "description": "服务器响应的时间戳",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/teacher/course_add": {
            "post": {
                "tags": [
                    "课程管理"
                ],
                "summary": "新增课程",
                "operationId": "7e0e2ac879d9c8266cf2ba1501b8a29d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_name",
                                    "cost",
                                    "date"
                                ],
                                "properties": {
                                    "course_name": {
                                        "description": "课程名",
                                        "type": "string"
                                    },
                                    "cost": {
                                        "description": "费用",
                                        "type": "string"
                                    },
                                    "date": {
                                        "description": "日期",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "返回码",
                                            "type": "integer"
                                        },
                                        "message": {
                                            "description": "错误信息",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "返回数据",
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "description": "服务器响应的时间戳",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Authorization-Bearer": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    }
}